Move scripts and styles back to subfolders
Move scripts back to scripts/ and styles back to
styles/. The problem is with Replit: for some
reason files created by the shell (including by
git or mv) will not be served until interacted
with in Replit's built-in editor. Typing and
deleting a single character in each new file
causes Replit to recognize it. This platform is
weird.